fix(cli): restore gateway-process test-isolation seam as a re-export barrel
The earlier dedup deleted gw-state-helpers.ts as a "100% duplicate", but it was
a load-bearing test-isolation seam, not accidental copy-paste. test/helpers/
cli-mocks.ts mock.module()s "../../src/lib/gateway-process.ts" PROCESS-GLOBALLY,
so once gateway-process.test.ts imported that same path, the combined
`bun test packages/cli/src` run fed it the leaked mock — 16 failures that pass
in isolation.

Keep the dedup (no copy-paste) AND the isolation: gw-state-helpers.ts now holds
the single real implementation; gateway-process.ts is a one-line re-export
barrel. Production still imports gateway-process.ts (still fully mockable); the
unit test imports gw-state-helpers.ts (the real, never-mocked module).
Verified: `bun test packages/cli/src` 999 pass / 0 fail; CLI tsc clean.
